+++ interface/libs/commands/include/cmd/spi_write/command.hxx
@@ -16,7 +16,7 @@ public:
 
         m_payload.at(offsetof(COMMANDS::SPI_WRITE::command_t, length)) = length;
 
-        for (int i = 0; i < sizeof(command.data); i++) {
+        for (uint8_t i = 0; i < sizeof(command.data); i++) {
             if (i >= data.size()) {
                 break;
             }
@@ -38,7 +38,7 @@ public:
 
         m_payload.at(offsetof(COMMANDS::SPI_WRITE::command_t, length)) = length;
 
-        for (int i = 0; i < data.size() && i < 32; i++) {
+        for (uint8_t i = 0; i < data.size() && i < 32; i++) {
             m_payload.at(
                 offsetof(COMMANDS::SPI_WRITE::command_t, data[0]) + i)
                 = data.at(i);
