+++ interface/libs/commands/include/cmd/ssd1306/command.hxx
@@ -14,7 +14,7 @@ public:
 
         m_payload.at(offsetof(COMMANDS::SSD1306::command_t, line)) = line;
 
-        for (int i = 0; i < sizeof(command.data); i++) {
+        for (uint8_t i = 0; i < sizeof(command.data); i++) {
             if (i >= data.size()) {
                 break;
             }
@@ -34,7 +34,7 @@ public:
 
         m_payload.at(offsetof(COMMANDS::SSD1306::command_t, line)) = line;
 
-        for (int i = 0; i < data.size() && i < 16; i++) {
+        for (uint8_t i = 0; i < data.size() && i < 16; i++) {
             m_payload.at(
                 offsetof(COMMANDS::SSD1306::command_t, data[0]) + i)
                 = data.at(i);
